[fix] tentative patch for missing refresh, explanation in description
When program comes to max hour refresh like 11pm, a delay are defined to restart the loop when we come to the min hour refresh like 6am. But I don't know why, the program do not restart. So I try this patch to solve the issue, but we've got battery issue so i can't test it now.